refactor: move src/http/openapi.ts to src/adapters/http/
