+++ src/adapters/http/openapi.ts
